Move test files in their specified folder
#80
