+++ backend/src/test/java/com/crave/backend/unit/dto/CommentDTOTest.java
@@ -1,4 +1,4 @@
-package com.crave.backend.unit;
+package com.crave.backend.unit.dto;
 
 import com.crave.backend.dto.CommentDTO;
 import com.crave.backend.enums.UserRole;
