+++ backend/src/test/java/com/crave/backend/unit/model/AccountTest.java
@@ -1,4 +1,4 @@
-package com.crave.backend.unit;
+package com.crave.backend.unit.model;
 
 import com.crave.backend.enums.UserRole;
 import com.crave.backend.model.Account;
