+++ backend/src/test/java/com/crave/backend/unit/model/RestaurantItemTest.java
@@ -1,4 +1,4 @@
-package com.crave.backend.unit;
+package com.crave.backend.unit.model;
 
 import com.crave.backend.model.RestaurantItem;
 import org.junit.jupiter.api.BeforeEach;
