+++ backend/src/test/java/com/crave/backend/unit/repository/CommentRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.crave.backend.unit;
+package com.crave.backend.unit.repository;
 
 import com.crave.backend.model.Comment;
 import com.crave.backend.model.Dish;
