+++ backend/src/test/java/com/crave/backend/unit/repository/TokenRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.crave.backend.unit;
+package com.crave.backend.unit.repository;
 
 import com.crave.backend.enums.TokenType;
 import com.crave.backend.enums.UserRole;
